Added new features
Added folder creation once experiment is run, detects if folder is created and if not, creates a folder under current date.
Inside the current date folder, another folder named linear_scan + the current time file was executed.

Data folder and plot folder is created under that folder as well.

updates in the code also embedded directory change, so the results would go to the folder they need to.